remove hocs
