+++ src/lib/Components/FillHandleRangeSelection.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { ProPaneContentChild } from './HOCs';
+import { PaneContentChild } from '../Model/InternalModel';
 
-export const FillHandleRangeSelection: React.FC<ProPaneContentChild> = ({ state, calculatedRange }) => {
+export const FillHandleRangeSelection: React.FC<PaneContentChild> = ({ state, calculatedRange }) => {
     return <>
         {state.currentBehavior.renderPanePart(state, calculatedRange)}
     </>
